requirements.txt
torch
torchvision
dominate
visdom
numpy
pillow
scikit-image
requests
scipy
opencv-python
tqdm